Repository

+++ src/main/java/pl/repository/CostRepository.java
@@ -3,7 +3,7 @@ package pl.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.entity.Cost;
 
-public interface CostRepository extends JpaRepository<Cost, Long> {
+public interface CostRepository extends JpaRepository<Cost, Integer> {
